move scaling to ScalableView
