+++ Wire-iOS/Sources/UserInterface/Calling/VideoGridView/VideoPreviews/VideoPreviewView.swift
@@ -21,7 +21,7 @@ import UIKit
 import avs
 import WireSyncEngine
 
-final class VideoPreviewView: BaseVideoPreviewView, UIGestureRecognizerDelegate {
+final class VideoPreviewView: BaseVideoPreviewView {
 
     var isPaused = false {
         didSet {
@@ -30,7 +30,6 @@ final class VideoPreviewView: BaseVideoPreviewView, UIGestureRecognizerDelegate 
         }
     }
 
-    private var previewContainer: UIView?
     private var previewView: AVSVideoView?
     private let blurView = UIVisualEffectView(effect: UIBlurEffect(style: .dark))
     private let pausedLabel = UILabel(
@@ -41,17 +40,15 @@ final class VideoPreviewView: BaseVideoPreviewView, UIGestureRecognizerDelegate 
         variant: .dark
     )
     private var snapshotView: UIView?
-    let pinchGesture = UIPinchGestureRecognizer()
-    let panGesture = UIPanGestureRecognizer()
 
     // MARK: - Initialization
 
-    init(stream: Stream, isCovered: Bool, shouldShowActiveSpeakerFrame: Bool, pinchToZoomRule: PinchToZoomRule) {
-        self.pinchToZoomRule = pinchToZoomRule
+    override init(stream: Stream, isCovered: Bool, shouldShowActiveSpeakerFrame: Bool, pinchToZoomRule: PinchToZoomRule) {
         super.init(
             stream: stream,
             isCovered: isCovered,
-            shouldShowActiveSpeakerFrame: shouldShowActiveSpeakerFrame
+            shouldShowActiveSpeakerFrame: shouldShowActiveSpeakerFrame,
+            pinchToZoomRule: pinchToZoomRule
         )
 
         updateState()
@@ -66,16 +63,6 @@ final class VideoPreviewView: BaseVideoPreviewView, UIGestureRecognizerDelegate 
             insertSubview($0, belowSubview: userDetailsView)
         }
         pausedLabel.textAlignment = .center
-
-        pinchGesture.addTarget(self, action: #selector(handlePinchGesture(_:)))
-        pinchGesture.delegate = self
-
-        panGesture.addTarget(self, action: #selector(handlePanGesture(_:)))
-        panGesture.delegate = self
-        panGesture.minimumNumberOfTouches = 2
-        panGesture.maximumNumberOfTouches = 2
-
-        updateGestureRecognizers()
     }
 
     override func createConstraints() {
@@ -85,85 +72,6 @@ final class VideoPreviewView: BaseVideoPreviewView, UIGestureRecognizerDelegate 
         pausedLabel.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
     }
 
-    // MARK: - Video scaling & panning
-
-    @objc
-    func handlePinchGesture(_ gestureRecognizer: UIPinchGestureRecognizer) {
-        guard
-            gestureRecognizer.state == .began
-            || gestureRecognizer.state == .changed
-            || gestureRecognizer.state == .ended
-        else { return }
-
-        guard let view = gestureRecognizer.view else { return }
-
-        // get location of the gesture's centroid
-        var location = gestureRecognizer.location(in: view)
-
-        // offset location relative to center of the view
-        // because transform is done relatively to the view's center
-        location.x -= view.bounds.midX
-        location.y -= view.bounds.midY
-
-        // translate view origin to pinch location, scale, translate view back
-        let transform = view.transform
-            .translatedBy(x: location.x, y: location.y)
-            .scaledBy(x: gestureRecognizer.scale, y: gestureRecognizer.scale)
-            .translatedBy(x: -location.x, y: -location.y)
-
-        // apply transform
-        view.transform = transform
-
-        // reset scale
-        gestureRecognizer.scale = 1
-
-        // reset to identity if the view is scaled smaller than its container
-        if view.frame.size.width < bounds.width {
-            view.transform = .identity
-        }
-    }
-
-    @objc
-    func handlePanGesture(_ gestureRecognizer: UIPanGestureRecognizer) {
-        guard let view = gestureRecognizer.view else { return }
-
-        // prevent translation if the view transform is identity
-        guard view.transform != .identity else { return }
-
-        let translation = gestureRecognizer.translation(in: view)
-
-        // translate view to gesture's location
-        view.transform = view.transform.translatedBy(x: translation.x, y: translation.y)
-
-        // reset translation
-        gestureRecognizer.setTranslation(.zero, in: view)
-    }
-
-    func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
-        return true
-    }
-
-    var pinchToZoomRule: PinchToZoomRule {
-        didSet {
-            guard oldValue != pinchToZoomRule else { return }
-            updateGestureRecognizers()
-        }
-    }
-
-    override func updateGestureRecognizers() {
-        let enabled = shouldEnableGestureRecognizers
-        panGesture.isEnabled = enabled
-        pinchGesture.isEnabled = enabled
-    }
-
-    private var shouldEnableGestureRecognizers: Bool {
-        switch pinchToZoomRule {
-        case .enableWhenFitted:
-            return !shouldFill
-        case .enableWhenMaximized:
-            return isMaximized
-        }
-    }
 
     // MARK: - Fill mode
 
@@ -235,19 +143,18 @@ final class VideoPreviewView: BaseVideoPreviewView, UIGestureRecognizerDelegate 
         previewView = preview
 
         // Adding the preview into a container allows smoother scaling
-        let container = UIView()
-        container.addSubview(preview)
-        container.addGestureRecognizer(pinchGesture)
-        container.addGestureRecognizer(panGesture)
-        previewContainer = container
+        let scalableView = ScalableView(isScalingEnabled: shouldEnableScaling)
+        scalableView.addSubview(preview)
+        self.scalableView?.removeFromSuperview()
+        self.scalableView = scalableView
 
         if let snapshotView = snapshotView {
-            insertSubview(container, belowSubview: snapshotView)
+            insertSubview(scalableView, belowSubview: snapshotView)
         } else {
-            insertSubview(container, belowSubview: userDetailsView)
+            insertSubview(scalableView, belowSubview: userDetailsView)
         }
 
-        [container, preview].forEach {
+        [scalableView, preview].forEach {
             $0.translatesAutoresizingMaskIntoConstraints = false
             $0.fitInSuperview()
         }
